src/requirements.txt
asttokens==2.4.1
attrs==23.2.0
bokeh==3.5.0
certifi==2024.7.4
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
hopcroftkarp==1.2.5
idna==3.7
igraph==0.11.5
imageio==2.34.2
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
persim==0.3.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyflagser==0.4.7
Pygments==2.18.0
pyparsing==3.1.2
-e git+ssh://git@github.com/aerojam95/rivet-python.git@7cd2d5eeffcb2913286359e9624f68899f0e5bc1#egg=pyrivet&subdirectory=../../../src/pyrivet
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
ripser==0.6.8
rpds-py==0.18.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.1
tenacity==8.3.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.7.24
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.6.0
